in spirit of DRY, reuse code, 2
